try to fix #45

+++ danmaku2ass.py
@@ -182,7 +182,7 @@ def ReadCommentsBilibili(f, fontsize):
             if comment.childNodes.length > 0:
                 if p[1] in ('1', '4', '5', '6'):
                     c = str(comment.childNodes[0].wholeText).replace('/n', '\n')
-                    size = int(p[2]) * fontsize / 25.0
+                    size = float(p[2]) * fontsize / 25.0
                     yield (float(p[0]), int(p[4]), i, c, {'1': 0, '4': 2, '5': 1, '6': 3}[p[1]], int(p[3]), size, (c.count('\n') + 1) * size, CalculateLength(c) * size)
                 elif p[1] == '7':  # positioned comment
                     c = str(comment.childNodes[0].wholeText)
